+++ Src/PerformanceCollector/Perf.NetFull.Tests/QuickPulse/QuickPulseTopCpuCollectorTests.cs
@@ -174,18 +174,15 @@
             var timeProvider = new ClockMock();
             var collector = new QuickPulseTopCpuCollector(timeProvider, processProvider);
 
-            var processDictionary =
-                QuickPulseTestHelper.GetPrivateField(collector, "processObservations") as Dictionary<string, TimeSpan>;
-
             // ACT
             collector.GetTopProcessesByCpu(3);
-            int itemCount1 = processDictionary.Count;
+            int itemCount1 = collector.ProcessObservations.Count;
 
             timeProvider.FastForward(TimeSpan.FromSeconds(1));
 
             processProvider.Processes = new List<QuickPulseProcess>() { new QuickPulseProcess("Process1", baseProcessorTime) };
             collector.GetTopProcessesByCpu(3);
-            int itemCount3 = processDictionary.Count;
+            int itemCount3 = collector.ProcessObservations.Count;
 
             // ASSERT
             Assert.AreEqual(5, itemCount1);