+++ Src/PerformanceCollector/Perf.Shared.Tests/QuickPulse/QuickPulseTestHelper.cs
@@ -1,13 +1,11 @@
 ﻿namespace Microsoft.ApplicationInsights.Tests
 {
     using System;
-    using System.Collections.Generic;
 #if NETCORE
     using System.Reflection;
 #endif
     using Microsoft.ApplicationInsights.Extensibility;
     using Microsoft.ApplicationInsights.Extensibility.Implementation;
-    using Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.QuickPulse;
 #if !NETCORE
     using Microsoft.VisualStudio.TestTools.UnitTesting;
 #endif
@@ -17,27 +15,10 @@
     /// </summary>
     internal static class QuickPulseTestHelper
     {
-        public static LinkedList<IQuickPulseTelemetryProcessor> GetTelemetryProcessors(QuickPulseTelemetryModule module)
-        {
-            return GetPrivateField(module, "telemetryProcessors") as LinkedList<IQuickPulseTelemetryProcessor>;
-        }
-
         public static void ClearEnvironment()
         {
             SetPrivateStaticField(typeof(TelemetryConfiguration), "active", null);
             TelemetryModules.Instance.Modules.Clear();
-        }
-
-        public static object GetPrivateField(object obj, string fieldName)
-        {
-#if NETCORE
-            TypeInfo type = obj.GetType().GetTypeInfo();
-            FieldInfo field = type.GetField(fieldName, BindingFlags.Instance | BindingFlags.NonPublic);
-            return field.GetValue(obj);
-#else
-            var t = new PrivateObject(obj);
-            return t.GetField(fieldName);
-#endif
         }
 
         private static void SetPrivateStaticField(Type type, string fieldName, object value)